+++ woofy-backend/src/main/java/com/woofy/woofy_backend/Repositories/CustomerRepository.java
@@ -3,5 +3,5 @@ package com.woofy.woofy_backend.Repositories;
 import com.woofy.woofy_backend.Models.Entities.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<CustomerEntity, Long> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity, Integer> {
 }
