+++ woofy-backend/src/main/java/com/woofy/woofy_backend/Repositories/DogRepository.java
@@ -3,4 +3,4 @@ package com.woofy.woofy_backend.Repositories;
 import com.woofy.woofy_backend.Models.Entities.DogEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DogRepository extends JpaRepository<DogEntity, Long> {}
+public interface DogRepository extends JpaRepository<DogEntity, Integer> {}
